doc: memfault: update README
Update of the requirement section to have memfault account and project
key.

Removed configuration files section as this one was not relevant for
the sample. All of these configuration files are already added to
the sample. When disabling the CONFIG_MEMFAULT_USER_CONFIG_ENABLE
it won't build the sample.

Added the option of connecting to the serial terminal through the
nrf connect VS Code extention serial terminal.

Signed-off-by: Kaja Koren <kaja.koren@nordicsemi.no>